simple-GAN/requirements.txt
imgaug==0.4.0
matplotlib==3.10.0
numpy==2.2.3
Pillow==11.1.0
PyYAML>=6.0
seaborn==0.13
torch==2.4
torchvision==0.19
tqdm==4.66
